+++ CHANGELOG.md
@@ -8,6 +8,49 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Released]
 
+## [0.2.0] - 2021-01-07
+### Changed
+- Bump version to 0.2.0
+
+## [0.1.9] - 2020-11-23
+### Added
+- Start tmux session
+
+## [0.1.8] - 2020-11-19
+### Changed
+- unset VAR
+
+## [0.1.7] - 2020-10-29
+### Added
+- Hello World
+
+## [0.1.6] - 2020-07-06
+### Changed
+- Redirect standard output to dev/null 
+- Renamed devnull.md to .sh
+
+## [0.1.5] - 2020-06-26
+### Changed
+- Display error msg when var is not set
+
+## [0.1.4] - 2020-04-17
+### Changed
+- String manipulation in bash update
+
+## [0.1.3] - 2020-04-12
+### Changed
+- String manipulation in bash update
+### Added
+- Add safeforweb
+
+## [0.1.2] - 2020-04-09
+### Added
+- Info on /dev/null usage
+
+## [0.1.1] - 2020-03-13
+### Added
+- String manipulation
+
 ## [0.1.0] - 2020-02-13
 ### Added
 - This CHANGELOG file